Added a delay to messagehandler
It was causing conflicts with fights because they both loaded/saved to database at the same time. A delay of 500ms for the messagehandling part should hopefully fix it. Also added "commands" as an alias for the help command, and removed the @here from negative pulsechecks (Will probably make the latter a server option sometime)