Music changes
The bot will now check once every minute if it's alone in the voice chat, and leave if so. Also added a property to ServerSpecifics that records the last channel used specifically for music commands, and the bot will send a message in that channel when leaving due to inactivity.