Various bugfixes
Fight and gamble no longer check for in progress because concurrency is hard. Top would crash the bot if you were lower rank than someone else who'd left the server, fixed by checking if user is null (maybe just replace with userId instead of a mention later?) CommandHandler line is not necessary yet, but if someone uses other bots to make infinite chains, I'll uncomment it.  Gaining exp in VC was broken because it returned after the first server not connected to VC (and the first server is my test server, so it's usually not connected there lmao), fixed by changing to continue