Certain commands no longer triggered by bots
Bot will no longer execute custom commands, admin commands, or setup commands if they're coming from a bot account, including itself. It used to just blanket respond to bots, but I realized that means users without perms can just make the bot use the kick/ban commands on itself, and it'd have permission to do so lol. Also added a line for it in the echo function, as it can technically be used to set up infinite chains together with other bots, but I'll let people have fun with it before I pull the trigger on that one.