Fixed a couple music bugs and twitterreplacer
Music module should now display a message when failing to load a song, and should also properly remove the player from voicechat, preventing it from getting stuck on the "you need to be in the same VC" error, also added a !join command just in case it still fails somehow. Twitterreplacer now always properly removes the tracker for platforms, and fixed a typo in the description when enabling delete mode.